+++ CHANGELOG.md
@@ -32,6 +32,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Allow CSS between empty `@media` rule and another `@media` rule
+  ([#534](https://github.com/MyIntervals/emogrifier/pull/534))
 - Allow additional whitespace in media-query-list of disallowed `@media` rules
   ([#532](https://github.com/MyIntervals/emogrifier/pull/532))
 - Allow multiple minified `@import` rules in the CSS without error (note: